requirements.txt

pandas
gspread
oauth2client
langchain_openai
beautifulsoup4
requests
